correcciones minimas en endpoint de registro de peso, servicios de productos y tiempo de validacion de JWT a 5hs

+++ src/weight_register/model/entities/weight_register.entity.ts
@@ -5,7 +5,7 @@ import { User } from "src/users/model/entities/user.entity";
 @Table
 export class WeightRegister extends Model {
 
-    @Column({ type: DataType.FLOAT(4, 2) })
+    @Column({ type: DataType.FLOAT(6, 2) })
     weight: number;
 
     @ForeignKey(() => Product)
